Struct tor_proto::channel::Channel

pub struct Channel { /* private fields */ }

An open client channel, ready to send and receive Tor cells.

A channel is a direct connection to a Tor relay, implemented using TLS.

This struct is a frontend that can be used to send cells (using the Sink<ChanCell> impl and otherwise control the channel. The main state is in the Reactor object. Channel is cheap to clone.

(Users need a mutable reference because of the types in Sink, and ultimately because cell_tx: mpsc::Sender doesn’t work without mut.

Channel life cycle

Channels can be created directly here through the ChannelBuilder API. For a higher-level API (with better support for TLS, pluggable transports, and channel reuse) see the tor-chanmgr crate.

After a channel is created, it will persist until it is closed in one of four ways:

A remote error occurs.

The other side of the channel closes the channel.

Someone calls Channel::terminate on the channel.

The last reference to the Channel is dropped. (Note that every circuit on a Channel keeps a reference to it, which will in turn keep the channel from closing until all those circuits have gone away.)

Note that in cases 1-3, the Channel object itself will still exist: it will just be unusable for most purposes. Most operations on it will fail with an error.

Implementations

impl Channel

pub fn unique_id(&self) -> UniqId

Return a process-unique identifier for this channel.

pub fn target(&self) -> &OwnedChanTarget

Return an OwnedChanTarget representing the actual handshake used to create this channel.

pub fn age(&self) -> Duration

Return the amount of time that has passed since this channel became open.

pub fn clock_skew(&self) -> ClockSkew

Return a ClockSkew declaring how much clock skew the other side of this channel claimed that we had when we negotiated the connection.

pub fn engage_padding_activities(&self)

Specify that this channel should do activities related to channel padding

Initially, the channel does nothing related to channel padding: it neither sends any padding, nor sends any PADDING_NEGOTIATE cells.

After this function has been called, it will do both, according to the parameters specified through reparameterize. Note that this might include disabling padding (for example, by sending a PADDING_NEGOTIATE).

Idempotent.

There is no way to undo the effect of this call.

pub fn reparameterize( &self, params: Arc<ChannelPaddingInstructionsUpdates> ) -> Result<()>

Reparameterise (update parameters; reconfigure)

Returns Err if the channel was closed earlier

pub fn check_match<T: HasRelayIds + ?Sized>(&self, target: &T) -> Result<()>

Return an error if this channel is somehow mismatched with the given target.

pub fn is_closing(&self) -> bool

Return true if this channel is closed and therefore unusable.

pub fn duration_unused(&self) -> Option<Duration>

If the channel is not in use, return the amount of time it has had with no circuits.

Return None if the channel is currently in use.

pub fn poll_ready(&mut self, cx: &mut Context<'_>) -> Result<bool>

Like futures::Sink::poll_ready.

pub async fn send_cell(&mut self, cell: AnyChanCell) -> Result<()>

Transmit a single cell on a channel.

pub async fn new_circ(&self) -> Result<(PendingClientCirc, Reactor)>

Return a newly allocated PendingClientCirc object with a corresponding circuit reactor. A circuit ID is allocated, but no messages are sent, and no cryptography is done.

To use the results of this method, call Reactor::run() in a new task, then use the methods of crate::circuit::PendingClientCirc to build the circuit.

pub fn terminate(&self)

Shut down this channel immediately, along with all circuits that are using it.

Note that other references to this channel may exist. If they do, they will stop working after you call this function.

It’s not necessary to call this method if you’re just done with a channel: the channel should close on its own once nothing is using it any more.

pub fn close_circuit(&self, circid: CircId) -> Result<()>

Tell the reactor that the circuit with the given ID has gone away.