Module tor_proto::stream

Implements Tor’s “stream“s from a client perspective

A stream is an anonymized conversation; multiple streams can be multiplexed over a single circuit.

To create a stream, use crate::circuit::ClientCirc::begin_stream.

Limitations

There is no fairness, rate-limiting, or flow control.

Structs

DataReader

The read half of a DataStream, implementing futures::io::AsyncRead.

DataStream

An anonymized stream over the Tor network.

DataWriter

The write half of a DataStream, implementing futures::io::AsyncWrite.

ResolveStream

A ResolveStream represents a pending DNS request made with a RESOLVE cell.

StreamParameters

A set of preferences used to declare how a new stream should be opened.

StreamReader

The read part of a stream on a particular circuit.

Enums

IpVersionPreference

A preference for IPv4 vs IPv6 addresses; usable as a nicer frontend for BeginFlags.